Start refactoring share view
